Fix gh#10: split media service module
Closes #10

Clawpatch finding: fnd_sig-gh-10-4fcb7008c2_a2f2811217
