+++ .clawpatch/features/feat_gh-10_f31d806dc5.json
@@ -1,0 +1,59 @@
+{
+  "schemaVersion": 1,
+  "featureId": "feat_gh-10_f31d806dc5",
+  "title": "gh#10: Refactor: Split media.service.ts god module",
+  "summary": "Imported from GitHub issue gh#10 in BACKLOG.md section \"Refactoring / type safety\".",
+  "kind": "service",
+  "source": "github-issue-import",
+  "confidence": "high",
+  "entrypoints": [
+    {
+      "path": "packages/api/src/services/media.service.ts",
+      "symbol": null,
+      "route": null,
+      "command": null
+    }
+  ],
+  "ownedFiles": [
+    {
+      "path": "packages/api/src/services/media.service.ts",
+      "reason": "mentioned by issue body"
+    }
+  ],
+  "contextFiles": [
+    {
+      "path": "BACKLOG.md",
+      "reason": "backlog queue item for gh#10"
+    }
+  ],
+  "tests": [],
+  "tags": [
+    "github-issue",
+    "gh#10",
+    "github-url:https://github.com/joshgk00/social-media-scheduler/issues/10",
+    "backlog-order:0047",
+    "backlog-section:refactoring-type-safety"
+  ],
+  "trustBoundaries": [
+    "network",
+    "filesystem"
+  ],
+  "status": "fixed",
+  "lock": null,
+  "findingIds": [
+    "fnd_sig-gh-10-4fcb7008c2_a2f2811217"
+  ],
+  "patchAttemptIds": [],
+  "analysisHistory": [
+    {
+      "runId": "20260527T040617-93c83b",
+      "kind": "github-issue-import",
+      "summary": "Imported GitHub issue gh#10",
+      "provider": "gh",
+      "model": null,
+      "createdAt": "2026-05-27T04:06:17.910Z"
+    }
+  ],
+  "createdAt": "2026-05-27T04:06:17.910Z",
+  "updatedAt": "2026-05-27T05:09:16.341Z"
+}
